Splitting up sfu.DownTrack
